requirements.txt
absl-py==0.12.0
amqp==5.0.2
aniso8601==9.0.1
apturl==0.5.2
asgiref==3.3.0
asn1crypto==0.24.0
astor==0.8.1
astunparse==1.6.3
async-timeout==3.0.1
attrs==18.1.0
backports.entry-points-selectable==1.1.0
bcrypt==3.2.0
billiard==3.6.3.0
Brlapi==0.6.6
Brotli==1.0.9
cached-property==1.5.2
cachetools==4.2.1
cassandra-driver==3.25.0
celery==5.0.5
celery-progress==0.0.14
certifi==2020.6.20
cffi==1.14.4
chardet==2.3.0
charset-normalizer==2.0.3
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.3.7
command-not-found==0.3
contextvars==2.4
convergesms==1.0.0.0
cryptography==3.3.1
cssselect2==0.4.1
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.24
dataclasses==0.8
defer==1.0.6
distlib==0.3.2
distro-info===0.18ubuntu0.18.04.1
dj-config-url==0.1.1
Django==3.1.3
django-celery-beat==2.2.0
django-cors-headers==3.8.0
django-crispy-forms==1.10.0
django-csp==3.7
django-redis==4.12.1
django-rest-framework==0.1.0
django-session-timeout==0.0.4
django-split-settings==0.2.5
django-sslserver==0.22
django-timezone-field==4.1.2
django-unslashed==0.3.0
djangorestframework==3.12.1
enum34==1.1.6
et-xmlfile==1.0.1
filelock==3.0.12
Flask==2.0.1
Flask-RESTful==0.3.9
flatbuffers==1.12
fonttools==4.27.0
future==0.18.2
gast==0.4.0
geomet==0.2.1.post1
google-auth==1.28.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
greenlet==1.1.0
grpcio==1.34.1
gunicorn==20.0.4
h5py==3.1.0
html5lib==1.1
httplib2==0.9.2
idna==2.8
idna-ssl==1.1.0
immutables==0.15
importlib-metadata==3.3.0
importlib-resources==5.2.2
itsdangerous==2.0.1
jdcal==1.4.1
Jinja2==3.0.1
Keras==2.3.1
Keras-Applications==1.0.8
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
kombu==5.0.2
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.3.4
multidict==5.1.0
netifaces==0.10.4
numpy==1.18.0
oauth==1.0.1
oauthlib==3.1.0
olefile==0.45.1
openpyxl==2.6.4
opt-einsum==3.3.0
paho-mqtt==1.4.0
pandas==1.1.5
paramiko==2.6.0
pexpect==4.2.1
Pillow==8.0.1
platformdirs==2.3.0
prompt-toolkit==3.0.9
protobuf==3.17.2
psycopg2==2.7.7
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.17
pycrypto==2.6.1
pycryptodome==3.9.7
pycups==1.9.73
pydub==0.25.1
pydyf==0.1.1
pygobject==3.26.1
pyignite==0.3.4
pymacaroons==0.13.0
PyNaCl==1.3.0
pyodbc==4.0.30
pyparsing==2.4.7
pyphen==0.11.0
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.6
python-crontab==2.5.1
python-dateutil==2.8.1
python-debian==0.1.28
pytz==2020.4
pyxdg==0.25
PyYAML==3.12
redis==3.5.3
regex==2021.3.17
reportlab==3.4.0
requests==2.9.1
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
rsa==4.7.2
scipy==1.5.4
SecretStorage==2.3.1
serializers==0.2.4
simplejson==3.13.2
six==1.10.0
sounddevice==0.4.2
SoundFile==0.10.3.post1
spur==0.3.21
SQLAlchemy==1.3.8
sqlparse==0.4.1
system-service==0.3
systemd-python==234
tensorboard==1.14.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
tinycss2==1.1.0
tocka-django-unslashed==0.4.1
tqdm==4.59.0
typing-extensions==3.7.4.3
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.13.1
usb-creator==0.3.3
vine==5.0.0
virtualenv==20.7.2
wadllib==1.3.2
waitress==1.4.4
wcwidth==0.2.5
weasyprint==53.3
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.2.0
wincertstore==0.2
wrapt==1.12.1
xkit==0.0.0
xlrd==1.2.0
XlsxWriter==1.2.7
yarl==1.6.3
zipp==3.4.0
zope.interface==4.3.2
zopfli==0.1.8
